Fix completions with empty values
Previously, completions with an empty string for `value` would display
correctly as a suggestion in the popup, but would error trying to
`Document.$detectNewLine` on the completion object if the user tried to
insert that suggestion.

Autocomplete now respects empty string as a value, doing an insert of
nothing on select that suggestion (rather than erroring).  This is handy
to properly support since it enables use cases like popping up a hint
in place of completions in cases where no suggestions are possible.