Fix autocomplete prefix for custom identifierRegexes
NOTE: this was sent upstream in https://github.com/ajaxorg/ace/pull/2352,
but closed because https://github.com/ajaxorg/ace/commit/fcebd0f662e0434fadf15faaaf4b59ab8950217a
seemed to address the same issue.  Our functional test do not like that
fix however, so we continue to use this one.

Autocomplete prefix can be wrong for completions with a custom
identifierRegex because `Autocompleter.base` is computed relative to a
prefix computed for the default identifierRegex.