Allow user feedback for empty autocompletes
Add support for setting a completer.emptyMessage function which
generates a message to display in the completions popup when there
are no completions.

This enables configuring the editor to always give the user feedback
when they ask for suggestions, even if it is just to note that there
are none.

This is a port of our earlier patch: https://github.com/looker/ace/commit/f328ef390405a2c60f7e64a1838370b97fe44062