Always wrap Ace output in a class so we can grab it with css
We have styling needs that Ace's default classes make impossible, so
we ensure all Ace-generated tokens are wrapped in the `ace_token` class
so we can target them.

Risk: Ace has added quite a few tests relying on the structure of the
output html, which is annoying to maintain, and may also indicate that
our new spans in particular could cause some problems at some point.

This is a port of our earlier patches:
https://github.com/looker/ace/commit/ba548b299a2007955fcbfeb61bb40d4061806d2e
https://github.com/looker/ace/commit/c0850e4013c358cafdaeea1b2701f167436189ce
https://github.com/looker/ace/commit/6721a7bbb2c9a2e6e8cad20e8038b354397e2be9