Append tooltip to body to ensure good positioning
Tooltip positioning was being messed up in our modal because of the
position on the modal's container.  Skip the issue by appending the
tooltip to the body instead of the editor container so that the mouse
coords it uses to compute position are always correct.

This is a port of our earlier patch: https://github.com/looker/ace/commit/dc16326c38c4fb563fa9d445384f850995c5358b